Fix the delete room bug when a room was just created
Fix the bug that when delete a room which was just created and haven't reloaded yet, the jeUser hosting the room will remain in database.
That is because onCreateRoom subscription will not get the complete information including the host id when a room created.
change the step to create room solve this problem.